# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Spend",
    "description": "API для управления расходами",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/group/{group_id}": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get Group",
        "operationId": "get_group_api_v1_group__group_id__get",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          },
          {
            "name": "id_user",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_GroupResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/group/user/{user_id}/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get Users Groups",
        "operationId": "get_users_groups_api_v1_group_user__user_id__groups_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_UserGroupsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/group/create": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Create Group",
        "operationId": "create_group_api_v1_group_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_GroupResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/group/delete": {
      "delete": {
        "tags": [
          "groups"
        ],
        "summary": "Delete Group",
        "operationId": "delete_group_api_v1_group_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Group Api V1 Group Delete Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Детали ошибки"
      },
      "GroupCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreate"
      },
      "GroupDelete": {
        "properties": {
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          }
        },
        "type": "object",
        "required": [
          "group_id"
        ],
        "title": "GroupDelete"
      },
...
      "GroupResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/UserRead"
            },
            "type": "array",
            "title": "Members",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "GroupResponse",
        "description": "Информация о группе с её участниками\n\n    id (int): ID группы\n    members (list): Список пользователей, состоящих в группе"
      },
      "GroupShort": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "GroupShort",
        "description": "Краткая информация о группе (для списков групп пользователя)\n\n    id (int): ID группы\n    name (str): Название группы"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StandardResponse_GroupResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Успешность выполнения запроса"
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "HTTP код ответа"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Данные ответа"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Информация об ошибке"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Временная метка"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code"
        ],
        "title": "StandardResponse[GroupResponse]",
        "example": {
          "code": 200,
          "data": {
            "user": {
              "id": 1,
              "name": "Nikita"
            }
          },
          "success": true,
          "timestamp": "2025-11-21T14:12:00Z"
        }
      },
...
      "StandardResponse_UserGroupsResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Успешность выполнения запроса"
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "HTTP код ответа"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserGroupsResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Данные ответа"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Информация об ошибке"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Временная метка"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code"
        ],
        "title": "StandardResponse[UserGroupsResponse]",
        "example": {
          "code": 200,
          "data": {
            "user": {
              "id": 1,
              "name": "Nikita"
            }
          },
          "success": true,
          "timestamp": "2025-11-21T14:12:00Z"
        }
      },
...
      "UserGroupsResponse": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupShort"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "groups"
        ],
        "title": "UserGroupsResponse",
        "description": "Список групп, в которых состоит пользователь\n\n        groups (list)"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}